+++ src/runtime/components/content/ContentSearch.vue
@@ -5,7 +5,7 @@ import type { ContentNavigationItem } from '@nuxt/content'
 import type { AppConfig } from '@nuxt/schema'
 import type { UseFuseOptions } from '@vueuse/integrations/useFuse'
 import theme from '#build/b24ui/content/content-search'
-import type { ButtonProps, InputProps, LinkProps, ModalProps, CommandPaletteProps, CommandPaletteSlots, CommandPaletteGroup, CommandPaletteItem, IconComponent, LinkPropsKeys } from '../../types'
+import type { ButtonProps, LinkProps, ModalProps, CommandPaletteProps, CommandPaletteSlots, CommandPaletteGroup, CommandPaletteItem, IconComponent, LinkPropsKeys } from '../../types'
 import type { ComponentConfig } from '../../types/tv'
 
 type ContentSearch = ComponentConfig<typeof theme, AppConfig, 'contentSearch'>
@@ -39,29 +39,11 @@ export interface ContentSearchItem extends Omit<LinkProps, 'custom'>, CommandPal
 /**
  * @memo not use loadingIcon
  */
-export interface ContentSearchProps<T extends ContentSearchLink = ContentSearchLink> extends Pick<ModalProps, 'title' | 'description' | 'overlay' | 'transition' | 'content' | 'dismissible' | 'fullscreen' | 'modal' | 'portal'> {
+export interface ContentSearchProps<T extends ContentSearchLink = ContentSearchLink> extends Pick<ModalProps, 'title' | 'description' | 'overlay' | 'transition' | 'content' | 'dismissible' | 'fullscreen' | 'modal' | 'portal'>, Pick<CommandPaletteProps<CommandPaletteGroup<ContentSearchItem>, ContentSearchItem>, 'icon' | 'placeholder' | 'autofocus' | 'loading' | 'closeIcon' | 'groups'> {
   /**
    * @defaultValue 'md'
    */
   size?: ContentSearch['variants']['size']
-  /**
-   * The icon displayed in the input.
-   * @defaultValue icons.search
-   * @IconComponent
-   */
-  icon?: IconComponent
-  /**
-   * The placeholder text for the input.
-   * @defaultValue t('commandPalette.placeholder')
-   */
-  placeholder?: InputProps['placeholder']
-  /**
-   * Automatically focus the input when component is mounted.
-   * @defaultValue true
-   */
-  autofocus?: boolean
-  /** When `true`, the loading icon will be displayed. */
-  loading?: boolean
   /**
    * Display a close button in the input (useful when inside a Modal for example).
    * `{ size: 'sm', color: 'air-tertiary-no-accent' }`{lang="ts-type"}
@@ -70,12 +52,6 @@ export interface ContentSearchProps<T extends ContentSearchLink = ContentSearchL
    */
   close?: boolean | Omit<ButtonProps, LinkPropsKeys>
   /**
-   * The icon displayed in the close button.
-   * @defaultValue icons.close
-   * @IconComponent
-   */
-  closeIcon?: IconComponent
-  /**
    * Keyboard shortcut to open the search (used by [`defineShortcuts`](https://bitrix24.github.io/b24ui/docs/composables/define-shortcuts/))
    * @defaultValue 'meta_k'
    */
@@ -83,12 +59,19 @@ export interface ContentSearchProps<T extends ContentSearchLink = ContentSearchL
   /** Links group displayed as the first group in the command palette. */
   links?: T[]
   navigation?: ContentNavigationItem[]
-  /** Custom groups displayed between navigation and color mode group. */
-  groups?: CommandPaletteGroup<ContentSearchItem>[]
   files?: ContentSearchFile[]
   /**
    * Options for [useFuse](https://vueuse.org/integrations/useFuse) passed to the [CommandPalette](https://bitrix24.github.io/b24ui/docs/components/command-palette/).
-   * @defaultValue { fuseOptions: { includeMatches: true } }
+   * @defaultValue  {
+      fuseOptions: {
+        ignoreLocation: true,
+        includeMatches: true,
+        threshold: 0.1,
+        keys: ['label', 'suffix']
+      },
+      resultLimit: 12,
+      matchAllWhenSearchEmpty: true
+    }
    */
   fuse?: UseFuseOptions<T>
   /**