+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.6.1](https://github.com/bitrix24/b24ui/compare/v2.6.0...v2.6.1) (2026-04-xx)
 
 * feat(CommandPalette): add `searchDelay` prop
+* fix(ContentSearch/DashboardSearch): pick shared props from CommandPalette
 
 
 ## [2.6.0](https://github.com/bitrix24/b24ui/compare/v2.5.3...v2.6.0) (2026-04-23)